+++ plugins/Goals/src/Events/GoalCreated.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace LifeWheel\Plugins\Goals\Events;
+
+use App\Models\User;
+
+final readonly class GoalCreated
+{
+    public function __construct(
+        public User $user,
+        public int $goalId,
+    ) {
+    }
+}