+++ plugins/Goals/src/Events/GoalMilestoneCompleted.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace LifeWheel\Plugins\Goals\Events;
+
+use App\Models\User;
+
+final readonly class GoalMilestoneCompleted
+{
+    public function __construct(
+        public User $user,
+        public int $goalId,
+        public int $milestoneId,
+    ) {
+    }
+}